% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpath.R
\docType{methods}
\name{dbConnect-dbpath}
\alias{dbConnect-dbpath}
\alias{dbConnect,dbpath-method}
\title{dbConnect method for dbpath}
\usage{
\S4method{dbConnect}{dbpath}(drv, ...)
}
\arguments{
\item{drv}{a driver instance}

\item{...}{extra arguments}
}
\description{
dbConnect method for dbpath
}
