% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpath.R
\name{dbpath}
\alias{dbpath}
\title{Produce an ast or parse graph of an expression}
\usage{
dbpath(url)
}
\arguments{
\item{url}{a database url}
}
\value{
A dbpath object
}
\description{
Produce an ast or parse graph of an expression
}
\examples{
dbpath("postgresql+psycopg2://some_user:some_pass@localhost:5432/some_db")

}
