% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpath_build.R
\name{dbpath_build}
\alias{dbpath_build}
\title{Build a dbpath object}
\usage{
dbpath_build(
  dialect,
  driver = NULL,
  username = NULL,
  password = NULL,
  host = NULL,
  port = NULL,
  database = NULL,
  params = NULL
)
}
\arguments{
\item{dialect, driver}{The SQL database \code{dialect} and optional R \code{driver}
package to use with this dialect.}

\item{username, password}{Optional username and password}

\item{host, port}{Optional host and port}

\item{database}{The name of the database}

\item{params}{A named list of query parameters or a query string. For
example, either \code{list(foo = "bar", name = "value")} or
\code{foo=bar&name=value}.}
}
\value{
A \code{\link[=dbpath]{dbpath()}} object.
}
\description{
Builds a \code{\link[=dbpath]{dbpath()}} object from its parts. All parts are optional other than
\code{dialect}. To create a database URL, use \code{format()} on the resulting \code{dbpath}
object.
}
\examples{
dbpath_build("postgres", host = "localhost", port = 5432, database = "money")

}
