% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{driver_defaults}
\alias{driver_defaults}
\title{configure default drivers for different dialects}
\format{
An object of class \code{list} of length 2.
}
\usage{
driver_defaults
}
\description{
configure default drivers for different dialects
}
\keyword{datasets}
