% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{driver_registry}
\alias{driver_registry}
\title{configure hooks for instantiating drivers}
\format{
An object of class \code{list} of length 2.
}
\usage{
driver_registry
}
\description{
configure hooks for instantiating drivers
}
\keyword{datasets}
