% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpath.R
\name{print.dbpath}
\alias{print.dbpath}
\title{Print a dbpath object}
\usage{
\method{print}{dbpath}(x, hide_password = TRUE, ..., url_encode = FALSE)
}
\arguments{
\item{x}{A \code{\link[=dbpath]{dbpath()}} object to print}

\item{hide_password}{Replace password with '****' if \link{TRUE}. Passwords are
hidden by default when printing a \code{\link[=dbpath]{dbpath()}} object, but are revealed when
using \code{format()} to construct a URL.}

\item{...}{extra arguments}

\item{url_encode}{If \link{TRUE}, the password and query paraemeters are
URL-encoded. Turned on by default in \code{format()}.}
}
\description{
Print a dbpath object
}
