% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpath.R
\name{format.dbpath}
\alias{format.dbpath}
\title{Format a dbpath object}
\usage{
\method{format}{dbpath}(x, hide_password = FALSE, ..., url_encode = TRUE)
}
\arguments{
\item{x}{The \code{\link[=dbpath]{dbpath()}} object}

\item{hide_password}{replace password with '****'}

\item{...}{extra arguments}

\item{url_encode}{If \link{TRUE}, the password and query paraemeters are
URL-encoded. Turned on by default in \code{format()}.}
}
\value{
A character string consisting of a dbpath URL, e.g
\verb{<dialect>+<driver>://<username>:<password>@<host>:<port>/<database>}.
}
\description{
Returns a formatted dbpath URL as a character string.
}
