% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpath.R
\name{print.dbpath}
\alias{print.dbpath}
\title{print a dbpath object}
\usage{
\method{print}{dbpath}(x, hide_password = TRUE, ...)
}
\arguments{
\item{x}{an item to print}

\item{hide_password}{replace password with '****'}

\item{...}{extra arguments}
}
\description{
print a dbpath object
}
